pyproject.toml
[tool.poetry]
name = "guardian"
version = "0.1.0"
description = "Multi-Cloud Workload Placement Optimizer Kubernetes Operator"
authors = ["Rajesh Ramesh <rramesh17993@gmail.com>"]
readme = "README.md"
packages = [{include = "guardian", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
kopf = "^1.37.0"
kubernetes = "^29.0.0"
scikit-learn = "^1.4.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
pydantic = "^2.6.0"
aiohttp = "^3.9.0"
xgboost = "^2.0.0"
prometheus-client = "^0.19.0"
boto3 = "^1.34.0"
azure-identity = "^1.15.0"
azure-mgmt-compute = "^30.0.0"
google-cloud-compute = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
